Force process_mapping to return balanced partitions.
Add cycle-refinement after initial partitioning to process_mapping.
The number of desired partitions to the kaffpa call is computed as the product of the hierarchy.
The recursive stopping condition in fast_construct_mapping::construct_initial_mapping_topdown_internal is adjusted to terminate properly even when
the number of paritions is not 1.